fix: fix bug in function to generate top Distances
The previous feature used 2 functions and  returned the arrays as single content arrays, I discovered that the array was declared within the for loop.

This has been fixed, and the getTopDistanceArray function returns an array of the distance of each section from the top of the viewport. The array length is the same as the number of sections.